Point Edit on GitHub at folder index.mdx files
Product landings live at content/docs/<product>/index.mdx. The TOC footer used to build content/docs/<product>.mdx, which does not exist. Resolve the source file on the server by trying the slug file first and falling back to index.mdx.
